Adjust names for correct use of ResourceDefinition and Resource. Also changes for i18n.
