+++ clustering/jgroups/src/main/java/org/jboss/as/clustering/jgroups/subsystem/JGroupsSubsystemXMLReader_1_1.java
@@ -61,7 +61,7 @@ public class JGroupsSubsystemXMLReader_1_1 implements XMLElementReader<List<Mode
             Attribute attribute = Attribute.forName(reader.getAttributeLocalName(i));
             switch (attribute) {
                 case DEFAULT_STACK: {
-                    JGroupsSubsystemRootResource.DEFAULT_STACK.parseAndSetParameter(value, subsystem, reader);
+                    JGroupsSubsystemRootResourceDefinition.DEFAULT_STACK.parseAndSetParameter(value, subsystem, reader);
                     break;
                 }
                 default: {
@@ -155,50 +155,50 @@ public class JGroupsSubsystemXMLReader_1_1 implements XMLElementReader<List<Mode
                 case TYPE: {
                     try {
                         TP.class.getClassLoader().loadClass(org.jgroups.conf.ProtocolConfiguration.protocol_prefix + '.' + value).asSubclass(TP.class).newInstance();
-                        TransportResource.TYPE.parseAndSetParameter(value, transport, reader);
+                        TransportResourceDefinition.TYPE.parseAndSetParameter(value, transport, reader);
                     } catch (Exception e) {
                         throw ParseUtils.invalidAttributeValue(reader, i);
                     }
                     break;
                 }
                 case SHARED: {
-                    TransportResource.SHARED.parseAndSetParameter(value, transport, reader);
+                    TransportResourceDefinition.SHARED.parseAndSetParameter(value, transport, reader);
                     break;
                 }
                 case SOCKET_BINDING: {
-                    TransportResource.SOCKET_BINDING.parseAndSetParameter(value, transport, reader);
+                    TransportResourceDefinition.SOCKET_BINDING.parseAndSetParameter(value, transport, reader);
                     break;
                 }
                 case DIAGNOSTICS_SOCKET_BINDING: {
-                    TransportResource.DIAGNOSTICS_SOCKET_BINDING.parseAndSetParameter(value, transport, reader);
+                    TransportResourceDefinition.DIAGNOSTICS_SOCKET_BINDING.parseAndSetParameter(value, transport, reader);
                     break;
                 }
                 case DEFAULT_EXECUTOR: {
-                    TransportResource.DEFAULT_EXECUTOR.parseAndSetParameter(value, transport, reader);
+                    TransportResourceDefinition.DEFAULT_EXECUTOR.parseAndSetParameter(value, transport, reader);
                     break;
                 }
                 case OOB_EXECUTOR: {
-                    TransportResource.OOB_EXECUTOR.parseAndSetParameter(value, transport, reader);
+                    TransportResourceDefinition.OOB_EXECUTOR.parseAndSetParameter(value, transport, reader);
                     break;
                 }
                 case TIMER_EXECUTOR: {
-                    TransportResource.TIMER_EXECUTOR.parseAndSetParameter(value, transport, reader);
+                    TransportResourceDefinition.TIMER_EXECUTOR.parseAndSetParameter(value, transport, reader);
                     break;
                 }
                 case THREAD_FACTORY: {
-                    TransportResource.THREAD_FACTORY.parseAndSetParameter(value, transport, reader);
+                    TransportResourceDefinition.THREAD_FACTORY.parseAndSetParameter(value, transport, reader);
                     break;
                 }
                 case SITE: {
-                    TransportResource.SITE.parseAndSetParameter(value, transport, reader);
+                    TransportResourceDefinition.SITE.parseAndSetParameter(value, transport, reader);
                     break;
                 }
                 case RACK: {
-                    TransportResource.RACK.parseAndSetParameter(value, transport, reader);
+                    TransportResourceDefinition.RACK.parseAndSetParameter(value, transport, reader);
                     break;
                 }
                 case MACHINE: {
-                    TransportResource.MACHINE.parseAndSetParameter(value, transport, reader);
+                    TransportResourceDefinition.MACHINE.parseAndSetParameter(value, transport, reader);
                     break;
                 }
                 default: {
@@ -237,14 +237,14 @@ public class JGroupsSubsystemXMLReader_1_1 implements XMLElementReader<List<Mode
                 case TYPE: {
                     try {
                         Protocol.class.getClassLoader().loadClass(org.jgroups.conf.ProtocolConfiguration.protocol_prefix + '.' + value).asSubclass(Protocol.class).newInstance();
-                        ProtocolResource.TYPE.parseAndSetParameter(value, protocol, reader);
+                        ProtocolResourceDefinition.TYPE.parseAndSetParameter(value, protocol, reader);
                     } catch (Exception e) {
                         throw ParseUtils.invalidAttributeValue(reader, i);
                     }
                     break;
                 }
                 case SOCKET_BINDING: {
-                    ProtocolResource.SOCKET_BINDING.parseAndSetParameter(value, protocol, reader);
+                    ProtocolResourceDefinition.SOCKET_BINDING.parseAndSetParameter(value, protocol, reader);
                     break;
                 }
                 default: {
@@ -300,7 +300,7 @@ public class JGroupsSubsystemXMLReader_1_1 implements XMLElementReader<List<Mode
         ModelNode property = Util.createAddOperation(propertyAddress);
 
         // assign the value
-        PropertyResource.VALUE.parseAndSetParameter(propertyValue, property, reader);
+        PropertyResourceDefinition.VALUE.parseAndSetParameter(propertyValue, property, reader);
 
         operations.add(property);
     }
