+++ clustering/jgroups/src/main/java/org/jboss/as/clustering/jgroups/subsystem/ProtocolLayerAdd.java
@@ -54,7 +54,7 @@ public class ProtocolLayerAdd implements OperationStepHandler {
         final ModelNode subModel = resource.getModel();
 
         // validate the protocol type to be added
-        ModelNode type = ProtocolResource.TYPE.validateOperation(operation);
+        ModelNode type = ProtocolResourceDefinition.TYPE.validateOperation(operation);
         PathElement protocolRelativePath = PathElement.pathElement(ModelKeys.PROTOCOL, type.asString());
 
         // if child resource already exists, throw OFE
@@ -95,7 +95,7 @@ public class ProtocolLayerAdd implements OperationStepHandler {
                     throw JGroupsMessages.MESSAGES.propertyNotDefined(property.getName(), protocolRelativePath.toString());
                 }
                 // set the value of the property
-                PropertyResource.VALUE.validateAndSet(value, param.getModel());
+                PropertyResourceDefinition.VALUE.validateAndSet(value, param.getModel());
             }
         }
         // This needs a reload
